Refactor functionality to qgis_bot
